subscription: fix action support for RPC callbacks
Actions are RPCs tied to a data node. See RFC 7950 for more details.

The RPC callback arguments have been changed to include the actual
operation path. For RPCs, it is the Xpath used when subscribing,
however, for actions, it is the full data path (see the docstring of
RpcCallbackType for more details).

To avoid tedious implementation of RPC/action callbacks, the output data
tree is already initialized by libsysrepo.so and only the actual output
leafs and containers need to be created. For that reason, callbacks
should now only return a dict with the output nodes (see the docstring
of RpcCallbackType, examples and tests).

In order to make this work, libyang>=1.5.0 is required.

Link: https://tools.ietf.org/html/rfc7950#section-7.15
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>