subscription: rework module change callbacks API
Because of asyncio support, we need to get the configuration and changes
*before* invoking the callbacks. Even if the callback does not need the
configuration at all, it will be converted to a python dict.

However, converting very large configuration trees to python dict takes
time. In order to scale with large configs, completely remove the config
parameter passed to module change callbacks. To make up for this, the
changes objects have been modified to use libyang trees instead of
sysrepo.Value objects.

This allows applications to update a local cache of the configuration in
the python dict format by using libyang.xpath_* data dict manipulation
functions. A unit test has been added to check that this actually works
as intended.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>