session: change rpc_send API
To be consistent with what the RPC/action subscription callbacks are
dealing with, change SysrepoSession.rpc_send to expect an RPC/action
data path (i.e. with list key values) and only the "input" subtree:

 sess.rpc_send("/my-module:my-rpc", {"foo": "bar"})
 sess.rpc_send("/my-module:cont/list[key='val']/my-action", {"x": 2})

If the "input" subtree is empty (i.e. no parameters), rpc_send should be
called with an empty dict:

 sess.rpc_send("/my-module:my-rpc", {})
 sess.rpc_send("/my-module:cont/list[key='val']/my-action", {})

The return value is also only the "output" subtree. If the RPC/action
returns nothing, the return value will be an empty dict.

See README and tests for more examples.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>