Rework logging configuration
Logging via a python CFFI callback can be problematic in some cases.
Indeed, the CFFI callback will be called for every message regardless
of any log level. It may have a negative performance impact. It is
therefore disabled by default.

Also, to be consistent with libsysrepo.so, all logging is disabled by
default.

It may be enabled on standard error, syslog or python logging
selectively with the new configure_logging function. See examples and
tests.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>