connection: allow to enable feature in module
Currently, feature can only be enabled at connection start. With this
patch, the feature can be enabled after the connection is created.

It allows to enable feature after all modules are loaded. Since the new
version of Sysrepo install module in a synchronous way, this function is
a must have.