callback: create unsafe mode to retrieve implicit session
A new API for sync callbacks is created, where the implicit session is
returned instead of the list of changes. This allows, in sync
callbacks, accessing all data (including changed nodes). However, when
callback returns to Sysrepo, the session is freed, so user callback
should never keep a reference on this implicit session.