tests: fix oper_sub_with_extra_info
The oper_sub_with_extra_info currently fails with this error:

> ======================================================================
> FAIL: test_oper_sub_with_extra_info (test_subs_oper.OperSubscriptionTest)
> ----------------------------------------------------------------------
>Traceback (most recent call last):
>   File "/root/tmp/sysrepo-python/tests/test_subs_oper.py", line 105, in test_oper_sub_with_extra_info
>     self.assertEqual(oper_data, {"state": {}})
> AssertionError: {'state': {'system': {}, 'network': {}}} != {'state': {}}
> - {'state': {'network': {}, 'system': {}}}
> + {'state': {}}

It is safe to assume that sysrepo changed its behavior in current
branch, in a commit that I did not try to identify. There are several
ways to fix this, the simplest being to set keep_empty_containers to
False, and expect empty datas. Indeed, what really matters in this test
is checking the extra infos (which is done in the callback).

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>